[FSK Editor] bugfix: dates not restored from settings
Created and modified dates were saved as long integers but then they
were loaded as integers from settings, failing to retrieve the values
and assigning always 0 (1970).

+++ de.bund.bfr.knime.fsklab.nodes/src/de/bund/bfr/knime/fsklab/nodes/editor/FskEditorNodeSettings.java
@@ -111,8 +111,8 @@ class FskEditorNodeSettings {
 				null : FskMetaData.Software.valueOf(softwareString);
 		metaData.referenceDescription = settings.getString("referenceDescription", "");
 		metaData.referenceDescriptionLink = settings.getString("referenceDescriptionLink", "");
-		metaData.createdDate = new Date(settings.getInt("createdDate", 0));
-		metaData.modifiedDate = new Date(settings.getInt("modifiedDate", 0));
+		metaData.createdDate = new Date(settings.getLong("createdDate", 0));
+		metaData.modifiedDate = new Date(settings.getLong("modifiedDate", 0));
 		metaData.notes = settings.getString("notes", "");
 		metaData.curated = settings.getBoolean("curated", false);
